Migrate to PEP 517 compliant build
Uses a PEP 517 compliant build setup to modernise the packaging of this
library and avoid having an executable setup.py when building the wheel.
This also bumps the Python dependency to a minimum of 3.8 aligning with
the currently supported Python versions and uses newer linting tools
that work with the latest versions of pytest.